refactor: decompose firewall, add internal plugin api
