+++ src/main/js/firewall/plugins/audit.js
@@ -1,0 +1,3 @@
+export const auditPlugin = ({rule, entry, options}) => {
+  return 'deny'
+}
