+++ src/main/js/index.d.ts
@@ -1,7 +1,11 @@
+type LetAsync<T> = T | Promise<T>
+
 type TApp = {
   start: () => Promise<void>
   stop: () => Promise<void>
 }
+
+type TLogger = typeof console
 
 type TServerConfig = {
   host?: string
@@ -19,8 +23,10 @@ type TServerConfig = {
   extend?: string
 }
 
+type TPolicy = 'allow' | 'deny' | 'warn'
+
 type TRule = {
-  policy: 'allow' | 'deny' | 'warn'
+  policy?: TPolicy
   name?: string | string[]
   org?: string | string[]
   dateRange?: [string, string]
@@ -28,9 +34,12 @@ type TRule = {
   version?: string,
   license?: string | string[]
   username?: string | string[],
-  filter?: (opts: Record<string, any>) => boolean | undefined | null
+  filter?: (entry: Record<string, any>) => LetAsync<boolean | undefined | null>
   extend?: string
+  plugin?: TPluginConfig
 }
+
+type TPluginConfig = string | [string, any] | TPlugin | [TPlugin, any]
 
 type TCacheConfig = {
   ttl: number
@@ -51,6 +60,16 @@ type TConfig = {
   server: TServerConfig | TServerConfig[]
   firewall: TFirewallConfig
   extend?: string
+}
+
+type TValidationContext = {
+  options: any,
+  rule: TRule,
+  entry: Record<string, any>
+}
+
+type TPlugin = {
+  (context: TValidationContext): LetAsync<TPolicy>
 }
 
 export function createApp(config: string | TConfig | TConfig[]): Promise<TApp>
